extract buildDateRange into dedicated file
https://coveord.atlassian.net/browse/KIT-320
