+++ packages/headless/src/controllers/facets/range-facet/date-facet/headless-date-facet.test.ts
@@ -2,7 +2,6 @@ import {
   DateFacet,
   buildDateFacet,
   DateFacetOptions,
-  buildDateRange,
 } from './headless-date-facet';
 import {
   MockEngine,
@@ -17,7 +16,6 @@ import {
 import {buildMockDateFacetValue} from '../../../../test/mock-date-facet-value';
 import {buildMockDateFacetResponse} from '../../../../test/mock-date-facet-response';
 import {SearchAppState} from '../../../../state/search-app-state';
-import {DateRangeRequest} from '../../../../features/facets/range-facets/date-facet-set/interfaces/request';
 import * as FacetIdDeterminor from '../../_common/facet-id-determinor';
 import {buildMockDateFacetRequest} from '../../../../test/mock-date-facet-request';
 
@@ -96,84 +94,5 @@ describe('date facet', () => {
     ];
 
     expect(dateFacet.state.values).toEqual(values);
-  });
-
-  it('#buildDateRange generates the correct value for a numeric input', () => {
-    const dateRange = buildDateRange({
-      start: 721386625000,
-      end: 752922625000,
-    });
-
-    const expectedValues: DateRangeRequest = {
-      start: '1992/11/10@09:10:25',
-      end: '1993/11/10@09:10:25',
-      endInclusive: false,
-      state: 'idle',
-    };
-
-    expect(dateRange).toEqual(expectedValues);
-  });
-
-  it('#buildDateRange generates the correct value for a js date input', () => {
-    const dateRange = buildDateRange({
-      start: new Date(721386625000),
-      end: new Date(752922625000),
-    });
-
-    const expectedValues: DateRangeRequest = {
-      start: '1992/11/10@09:10:25',
-      end: '1993/11/10@09:10:25',
-      endInclusive: false,
-      state: 'idle',
-    };
-
-    expect(dateRange).toEqual(expectedValues);
-  });
-
-  it('#buildDateRange generates the correct value for an iso 8601 string input', () => {
-    const dateRange = buildDateRange({
-      start: new Date(721386625000).toISOString(),
-      end: new Date(752922625000).toISOString(),
-    });
-
-    const expectedValues: DateRangeRequest = {
-      start: '1992/11/10@09:10:25',
-      end: '1993/11/10@09:10:25',
-      endInclusive: false,
-      state: 'idle',
-    };
-
-    expect(dateRange).toEqual(expectedValues);
-  });
-
-  it('#buildDateRange throws if the date can not be parsed', () => {
-    expect(() =>
-      buildDateRange({
-        start: 'NOT A DATE',
-        end: 'NOT A DATE',
-      })
-    ).toThrow(
-      `Could not parse the provided date, please provide a dateFormat string in the configuration options.\n
-       See https://day.js.org/docs/en/parse/string-format for more information.
-       `
-    );
-  });
-
-  it('#buildDateRange uses provided date format string', () => {
-    const dateFormat = 'MM-YYYY-DD@HH:mm:ss';
-    const dateRange = buildDateRange({
-      start: '11-1992-10@09:10:25',
-      end: '11-1993-10@09:10:25',
-      dateFormat,
-      useLocalTime: true,
-    });
-
-    const expectedValues: DateRangeRequest = {
-      start: '1992/11/10@09:10:25',
-      end: '1993/11/10@09:10:25',
-      endInclusive: false,
-      state: 'idle',
-    };
-    expect(dateRange).toEqual(expectedValues);
   });
 });
