+++ packages/headless/src/controllers/facets/range-facet/date-facet/headless-date-facet.ts
@@ -1,11 +1,4 @@
-import dayjs from 'dayjs';
-import utc from 'dayjs/plugin/utc';
-import customParseFormat from 'dayjs/plugin/customParseFormat';
-
-import {
-  DateRangeRequest,
-  DateFacetRequest,
-} from '../../../../features/facets/range-facets/date-facet-set/interfaces/request';
+import {DateFacetRequest} from '../../../../features/facets/range-facets/date-facet-set/interfaces/request';
 import {Engine} from '../../../../app/headless-engine';
 import {DateFacetRegistrationOptions} from '../../../../features/facets/range-facets/date-facet-set/interfaces/options';
 import {
@@ -26,45 +19,9 @@ import {
   dateFacetOptionsSchema,
 } from './headless-date-facet-options';
 import {determineFacetId} from '../../_common/facet-id-determinor';
+import {buildDateRange} from './date-range';
 
-dayjs.extend(utc);
-dayjs.extend(customParseFormat);
-
-type DateRangeOptions = Partial<Omit<DateRangeRequest, 'start' | 'end'>> & {
-  start: string | number | Date;
-  end: string | number | Date;
-  useLocalTime?: boolean;
-  dateFormat?: string;
-};
-
-export function buildDateRange(config: DateRangeOptions): DateRangeRequest {
-  const DATE_FORMAT = 'YYYY/MM/DD@HH:mm:ss';
-  const start = config.useLocalTime
-    ? dayjs(config.start, config.dateFormat).format(DATE_FORMAT)
-    : dayjs(config.start, config.dateFormat).utc().format(DATE_FORMAT);
-  const end = config.useLocalTime
-    ? dayjs(config.end, config.dateFormat).format(DATE_FORMAT)
-    : dayjs(config.end, config.dateFormat).utc().format(DATE_FORMAT);
-
-  if (start === 'Invalid Date' || end === 'Invalid Date') {
-    throw new Error(
-      `Could not parse the provided date, please provide a dateFormat string in the configuration options.\n
-       See https://day.js.org/docs/en/parse/string-format for more information.
-       `
-    );
-  }
-
-  const endInclusive = config.endInclusive ?? false;
-  const state = config.state ?? 'idle';
-
-  return {
-    start,
-    end,
-    endInclusive,
-    state,
-  };
-}
-
+export {buildDateRange};
 export {DateFacetOptions};
 export type DateFacetProps = {
   options: DateFacetOptions;
